moving genesis json to db package
