+++ beacon-chain/blockchain/service_test.go
@@ -160,7 +160,7 @@ func TestRunningChainService(t *testing.T) {
 	defer chainService.beaconDB.Close()
 
 	active := types.NewGenesisActiveState()
-	crystallized, err := types.NewGenesisCrystallizedState("")
+	crystallized, err := types.NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("Can't generate genesis state: %v", err)
 	}
@@ -245,7 +245,7 @@ func TestUpdateHeadEmpty(t *testing.T) {
 	defer chainService.beaconDB.Close()
 
 	active := types.NewGenesisActiveState()
-	crystallized, err := types.NewGenesisCrystallizedState("")
+	crystallized, err := types.NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("Can't generate genesis state: %v", err)
 	}
@@ -358,7 +358,7 @@ func TestUpdateHead(t *testing.T) {
 	defer chainService.beaconDB.Close()
 
 	active := types.NewGenesisActiveState()
-	crystallized, err := types.NewGenesisCrystallizedState("")
+	crystallized, err := types.NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("Can't generate genesis state: %v", err)
 	}
@@ -409,7 +409,7 @@ func TestProcessBlocksWithCorrectAttestations(t *testing.T) {
 	defer chainService.beaconDB.Close()
 
 	active := types.NewGenesisActiveState()
-	crystallized, err := types.NewGenesisCrystallizedState("")
+	crystallized, err := types.NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("Can't generate genesis state: %v", err)
 	}
