+++ beacon-chain/db/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/types:go_default_library",
+        "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/database:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
