+++ beacon-chain/types/active_state_test.go
@@ -170,7 +170,7 @@ func TestCalculateNewBlockHashes_DoesNotMutateData(t *testing.T) {
 
 func TestBlockVoteCacheNoAttestations(t *testing.T) {
 	aState := NewGenesisActiveState()
-	cState, err := NewGenesisCrystallizedState("")
+	cState, err := NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("failed to initialize crystallized state: %v", err)
 	}
@@ -188,7 +188,7 @@ func TestBlockVoteCacheNoAttestations(t *testing.T) {
 
 func TestBlockVoteCache(t *testing.T) {
 	aState := NewGenesisActiveState()
-	cState, err := NewGenesisCrystallizedState("")
+	cState, err := NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("failed to initialize crystallized state: %v", err)
 	}
@@ -232,7 +232,7 @@ func TestCalculateNewActiveState(t *testing.T) {
 		AncestorHashes: [][]byte{{}},
 	})
 
-	cState, err := NewGenesisCrystallizedState("")
+	cState, err := NewGenesisCrystallizedState(nil)
 	if err != nil {
 		t.Fatalf("failed to initialize genesis crystallized state: %v", err)
 	}
