+++ beacon-chain/types/block_test.go
@@ -75,7 +75,7 @@ func TestGenesisBlock(t *testing.T) {
 }
 
 func TestBlockValidity(t *testing.T) {
-	cState, err := NewGenesisCrystallizedState("")
+	cState, err := NewGenesisCrystallizedState(nil)
 
 	if err != nil {
 		t.Fatalf("failed to generate crystallized state: %v", err)
